Fixes "superclass mismatch for class Name" error in edge Rails
