+++ lib/gettext_i18n_rails/active_model/name.rb
@@ -1,0 +1,7 @@
+module ActiveModel
+  Name.class_eval do
+    def human(options={})
+      _(@klass.humanize_class_name)
+    end
+  end
+end
