--- spec/gettext_i18n_rails/active_model_spec.rb
@@ -1,16 +1,0 @@
-# encoding: utf-8
-require "spec_helper"
-
-describe ActiveModel::Name do
-  before do
-    FastGettext.current_cache = {}
-  end
-
-  describe 'human' do
-    it "is translated through FastGettext" do
-      name = ActiveModel::Name.new(CarSeat)
-      name.should_receive(:_).with('Car seat').and_return('Autositz')
-      name.human.should == 'Autositz'
-    end
-  end
-end if ActiveRecord::VERSION::MAJOR >= 3
